It looks like RandomMatrices is only used in this package to generate Haar distributed unitary matrices within the methods `randstate_haar` and `randunitary_haar`. This PR implements a slightly faster algorithm that roughly halves the number of allocations and gets rid of RandomMatrices as a dependency.

Before:
```
julia> for i in 10:10:50
          b = SpinBasis(i)
          @btime randunitary_haar($b)
       end
  25.125 μs (41 allocations: 64.75 KiB)
  88.458 μs (41 allocations: 233.47 KiB)
  192.958 μs (41 allocations: 479.50 KiB)
  415.917 μs (41 allocations: 813.44 KiB)
  638.834 μs (41 allocations: 1.21 MiB)
```
After:
```
julia> for i in 10:10:50
           b = SpinBasis(i)
           @btime randunitary_haar($b)
       end
  23.958 μs (23 allocations: 49.94 KiB)
  84.666 μs (23 allocations: 179.42 KiB)
  185.708 μs (23 allocations: 361.56 KiB)
  397.792 μs (23 allocations: 606.28 KiB)
  598.959 μs (23 allocations: 913.59 KiB)
```