@Krastanov 2024-11-24
huh... who would have thought that RandomMatrices is not already excessively optimized.

Do you mind also checking if `qr!` makes things even faster?

Either way, this is great -- it also removes the transitive dependency on a bunch more packages. Should be merged pretty soon