@apkille 2024-11-24
> huh... who would have thought that RandomMatrices is not already excessively optimized.

Yeah, I was surprised by this too. I was planning on using RandomMatrices for Gabs but was shocked by the number of allocations it was producing. It also seems like RandomMatrices has been dealing with a maintenance issue for quite a while...

> Do you mind also checking if `qr!` makes things even faster?

Checked same benchmark as before and `qr!` gives nearly the same exact speed with a few less allocations, so I made the change.

> Either way, this is great -- it also removes the transitive dependency on a bunch more packages. Should be merged pretty soon